+++ Source/CoAPnet/Protocol/BlockTransfer/CoapBlockTransferOptionValue.cs
@@ -4,7 +4,7 @@
     {
         public ushort Number { get; set; }
 
-        public byte Size { get; set; }
+        public ushort Size { get; set; }
 
         public bool HasFollowingBlocks { get; set; }
     }
